+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+v0.3.0 - 2019-06-XX
+-------------------
+
+* Remove flake8 v2 compatibility
+* Remove `--copyright-check` boolean argument, now always check
+* Change plugin registration so `select` does not need modifying
+* Update message to clarify message may be missing or malformed
+* Default min file size to 3 so empty or newline-only files pass
+
 v0.2.2 - 2018-10-10
 -------------------